...
git-svn-id: https://repo.acin.tuwien.ac.at/v4r/trunk/ros/hobbit_hydro_src@10055 a44b1c4d-ba22-49c9-afbd-3e401e1c4b3e

+++ emergency_detector/src/tools.cpp
@@ -132,7 +132,7 @@ unsigned short * copyDepth(unsigned short * source , unsigned int width , unsign
 
 
 
-unsigned int temperatureSensorSensesHuman(unsigned int tempDetected , unsigned int tempTimestamp , unsigned int frameTimestamp)
+unsigned int temperatureSensorSensesHuman(float tempDetected , unsigned int tempTimestamp , unsigned int frameTimestamp)
 {
  unsigned int temperatureFrameOffset = ABSDIFF(frameTimestamp,tempTimestamp);
   if (
